Use setuptools for installing python bindings
Summary:
cmake is mostly for installing C++ code. For installing python it is best to go
with setuptools as it has better understanding about python-modules,
python-scripts as well as dependencies.

For e.g. where to install python-binding is now decided based on system default
python version (in cmake we were specifying to install under python2.7 which
is very fragile).

NOTE: This will require changes in Terragraph/SOMA (which I plan to do once I
also work out OpenR bindings)

Reviewed By: mcallahan

Differential Revision: D5386489

fbshipit-source-id: e89e4304e507140611a139bfce6aee243b9a7454