remove usage of 'std::function<... noexcept>
Summary:
The declaration 'std::function<... noexcept> is improper; std::function's
operator () is specified as throwing an exception. In earlier versions of
C++ functions with a different noexcept specifier were considered the
same type, but this is no longer the case in C++17. This leads to build
errors when building code using this construct with platform007.

Fix by stripping meaningless noexcept specifier. Instances located with
simple regex:

  fbgr 'std::function<[^>]*noexcept'

Reviewed By: mcallahan

Differential Revision: D12840310

fbshipit-source-id: 4a11b7a6fa2736e37b2369894e848a3615008dfc