getdeps: indicate that folly depends on lz4
Summary:
Update the manifest file for folly to indicate a dependency on lz4.
folly does not require lz4 be available, but it will use it if it is found at
configure time.

getdeps is unfortunately not strict about providing projects only with the
dependencies that they require at build time.  This causes it to sometimes
make lz4 available to folly (if you are also building another project that
requires lz4), and sometimes not.  This ends up causing changes in
folly-config.h depending on which projects you are building, forcing all of
the folly sources to be recompiled.

In the future we perhaps should update getdeps to consistently only pass in
include directories for dependencies actually listed in the manifest file.
However, specifying that folly depends on lz4 also works to mitigate this
particular issue for now, and it is also generally desirable to build folly
with lz4 support.

Reviewed By: wez

Differential Revision: D21359995

fbshipit-source-id: aaf61671b7750d6c47e3613c732d220b3311b5ba