allow overriding version
Summary:
When packaging for Fedora, we want to match the version to the release tag,
since we want to enable shared libraries (the norm for distributions) and
want to make sure that dependent packages get rebuilt.

Allow overriding both the version and the so version, and also extend this
to the ``monitor_cpp2` library.

Reviewed By: saifhhasan

Differential Revision: D24911180

fbshipit-source-id: da65de6d680a6df81dbc5b1ee0a38d0f4344ae0f