feat: allow CMAKE_CXX_FLAGS to be overwritten
This may be helpful if one wants to disable some warning
that was previously added by the project itself which
may cause problem on different envs.
For example:

Clang might warn about certain code constructions that
g++ will now, this causing the code to not be compiled
on OS X. In order to avoid such issues allow the caller
to supress such warnings.